Sorted out the zooming and panning at different window sizes. Added documentation to the PTMWindow class and reorganised things a bit so they are more optimised and organised better.

+++ src/shaders/defaultFragmentShader.glsl
@@ -3,8 +3,8 @@
 
 uniform float lightX;
 uniform float lightY;
-uniform int imageHeight;
-uniform int imageWidth;
+uniform float imageHeight;
+uniform float imageWidth;
 
 uniform isampler2D rVals1;
 uniform isampler2D rVals2;
