+++ src/shaders/diffuseGainFragmentShader.glsl
@@ -3,8 +3,8 @@
 
 uniform float lightX;
 uniform float lightY;
-uniform int imageHeight;
-uniform int imageWidth;
+uniform float imageHeight;
+uniform float imageWidth;
 
 uniform float diffGain;
